chore: GetBackupEncryptionStatus moved from proxy to backup-manager
Refs: XRDDEV-2896
