--- src/security-server/system-test/src/intTest/resources/container-files/Dockerfile
@@ -1,5 +1,0 @@
-ARG BASE_IMAGE="xrd-securityserver-dev:latest"
-
-FROM $BASE_IMAGE
-
-COPY --chown=xroad:xroad etc /etc
