+++ src/security-server/system-test/src/intTest/resources/container-files/etc/xroad/conf.d/local.yaml
@@ -16,8 +16,6 @@ xroad:
   proxy:
     ssl-enabled: false # Can be removed once UI system tests configure a functional AUTH certificate
     health-check-port: 5558
-    backup-encryption-enabled: true
-    backup-encryption-keyids: [ "backup.key1@example.org", "backup.key2@example.org", "backup.key3@example.org" ]
     addon:
       op-monitor:
         enabled: true
@@ -41,3 +39,6 @@ xroad:
       socket-timeout-seconds: 1
     service:
       socket-timeout-seconds: 1
+  backup-manager:
+    backup-encryption-enabled: true
+    backup-encryption-keyids: [ "backup.key1@example.org", "backup.key2@example.org", "backup.key3@example.org" ]
